feat(back) implementación para que solo un usuario pueda ver, crear, actualizar y eliminar su propia playlist

+++ src/main/java/com/Profpost/service/impl/PlaylistServiceImpl.java
@@ -7,10 +7,14 @@ import com.Profpost.exception.ResourceNotFoundExcept;
 import com.Profpost.mapper.PlaylistMapper;
 import com.Profpost.model.entity.Playlist;
 import com.Profpost.model.entity.Publication;
+import com.Profpost.model.entity.User;
 import com.Profpost.repository.PlaylistRepository;
 import com.Profpost.repository.PublicationRepository;
+import com.Profpost.repository.UserRepository;
 import com.Profpost.service.PlaylistService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -23,6 +27,21 @@ public class PlaylistServiceImpl implements PlaylistService {
     private final PlaylistRepository playlistRepository;
     private final PublicationRepository publicationRepository;
     private final PlaylistMapper playlistMapper;
+    private final UserRepository userRepository;
+
+    @Transactional
+    @Override
+    public List<PlaylistDTO> findPlaylistsByUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String email = authentication.getName();
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+
+        List<Playlist> playlists = playlistRepository.findByUserId(user.getId());
+        return playlists.stream()
+                .map(playlistMapper::toDTO)
+                .toList();
+    }
 
     @Transactional(readOnly = true)
     @Override
@@ -36,13 +55,19 @@ public class PlaylistServiceImpl implements PlaylistService {
     @Transactional
     @Override
     public PlaylistDTO create(PlaylistDTO playlistDTO) {
-        playlistRepository.findByName(playlistDTO.getName())
-                        .ifPresent(existingPlaylist -> {
-                            throw new BadRequestException("Playlist con nombre " + playlistDTO.getName() + " ya existe");
-                        });
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String email = authentication.getName();
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+        playlistRepository.findByNameAndUserId(playlistDTO.getName(), user.getId())
+                .ifPresent(existingPlaylist -> {
+                    throw new BadRequestException("Ya existe una playlist con el nombre '" + playlistDTO.getName() + "' para este usuario.");
+                });
+
         Playlist playlist = playlistMapper.toEntity(playlistDTO);
+        playlist.setUser(user);
         playlist.setCreated_at(LocalDateTime.now());
-        playlist=playlistRepository.save(playlist);
+        playlist = playlistRepository.save(playlist);
         return playlistMapper.toDTO(playlist);
     }
 
@@ -58,24 +83,45 @@ public class PlaylistServiceImpl implements PlaylistService {
     @Override
     public PlaylistDTO update(Integer id, PlaylistDTO updatedPlaylistDTO) {
         Playlist playlist = playlistRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundExcept("La Playlist con ID " +id+ "no encontrada"));
+                .orElseThrow(() -> new ResourceNotFoundExcept("La Playlist con ID " + id + " no encontrada"));
 
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String email = authentication.getName();
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+
+        if (!playlist.getUser().getId().equals(user.getId())) {
+            throw new RuntimeException("No tienes permiso para actualizar esta playlist");
+        }
+
+        // Verificar si ya existe una playlist con el mismo nombre, pero diferente ID
         playlistRepository.findByName(updatedPlaylistDTO.getName())
-                        .filter(existingPlaylist -> !existingPlaylist.getName().equals(id))
-                        .ifPresent(existingPlaylist -> {throw new BadRequestException("Ya existe otra playlist con el mismo nombre");
-                        });
+                .filter(existingPlaylist -> !existingPlaylist.getId().equals(id))
+                .ifPresent(existingPlaylist -> {
+                    throw new BadRequestException("Ya existe otra playlist con el mismo nombre");
+                });
 
         playlist.setName(updatedPlaylistDTO.getName());
         playlist.setUpdated_at(LocalDateTime.now());
-        playlist=playlistRepository.save(playlist);
+        playlist = playlistRepository.save(playlist);
         return playlistMapper.toDTO(playlist);
     }
+
 
     @Transactional
     @Override
     public void delete(Integer id) {
         Playlist playlist = playlistRepository.findById(id)
-                        .orElseThrow(() -> new ResourceNotFoundExcept("Playlist con ID "+id+" no encontrada"));
+                .orElseThrow(() -> new ResourceNotFoundExcept("Playlist con ID " + id + " no encontrada"));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String email = authentication.getName();
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+
+        if (!playlist.getUser().getId().equals(user.getId())) {
+            throw new RuntimeException("No tienes permiso para eliminar esta playlist");
+        }
         playlistRepository.delete(playlist);
     }
